+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - do not add the same message in a row to input history
 - removed Aider run_prompt tool responses to reduce token usage
 - added options to manage TODO list items (add, edit, remove)
+- interrupt action cancels the current question
 
 ## [0.18.0]